Add files via upload
Employee Management System is based on a concept to store and generate all the records 
of the employees. This program is considered as a simple database of employees in an 
office, an organization where the user can store employees record easily as it is not timeconsuming. Here at first, the user has to pass through login system to get access then 
he/she can add employee’s data, view list of employees, modify and remove employee 
details. The whole project is designed in ‘C++’ language and different variables and strings 
have been used for the development of this project.And for storing employee details i have 
used the file handling concept. This project is easy to operate and understood by the users.
FEATURES:
1. Login system
2. Add Record
3. Display Record
4. Search Record
5. Modify Record
6. Delete Record
7. Search group
8. Show all Groups Deails